Exposed internal IPRange structs in ip.hpp
