+++ include/oxen/quic.hpp
@@ -3,6 +3,7 @@
 #include "quic/address.hpp"
 #include "quic/btstream.hpp"
 #include "quic/connection.hpp"
+#include "quic/connection_ids.hpp"
 #include "quic/context.hpp"
 #include "quic/crypto.hpp"
 #include "quic/datagram.hpp"
@@ -12,6 +13,8 @@
 #include "quic/format.hpp"
 #include "quic/formattable.hpp"
 #include "quic/gnutls_crypto.hpp"
+#include "quic/iochannel.hpp"
+#include "quic/ip.hpp"
 #include "quic/messages.hpp"
 #include "quic/network.hpp"
 #include "quic/opt.hpp"
